Move PostgresConfig to a common place
