+++ api-server/api-server-common/src/lib.rs
@@ -14,3 +14,38 @@
 // limitations under the License.
 
 pub mod storage;
+
+use clap::Parser;
+
+#[derive(Parser, Debug)]
+pub struct PostgresConfig {
+    /// Postgres host
+    #[clap(long, default_value = "localhost")]
+    pub postgres_host: String,
+
+    /// Postgres port
+    #[clap(long, default_value = "5432")]
+    pub postgres_port: u16,
+
+    /// Postgres user
+    #[clap(long, default_value = "postgres")]
+    pub postgres_user: String,
+
+    /// Postgres password
+    #[clap(long)]
+    pub postgres_password: Option<String>,
+
+    /// Postgres database
+    #[clap(long)]
+    pub postgres_database: Option<String>,
+
+    /// Postgres max connections
+    #[clap(long, default_value = "10")]
+    pub postgres_max_connections: u32,
+}
+
+impl Default for PostgresConfig {
+    fn default() -> Self {
+        Self::parse()
+    }
+}
